% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actogram.R
\name{plot_actogram}
\alias{plot_actogram}
\title{plot_actogram}
\usage{
plot_actogram(data, start_date = min(data$date), end_date = max(data$date))
}
\arguments{
\item{data}{the activity data to plot}

\item{start_date}{the start time}

\item{end_date}{the end time}
}
\value{
the plot
}
\description{
plots an actogram
}
