% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clock_lab.R
\name{read_clock_lab_file}
\alias{read_clock_lab_file}
\title{read_clock_lab_file}
\usage{
read_clock_lab_file(file_name)
}
\arguments{
\item{file_name}{the name of the file to read, should represent one mouse.
Should be the result of a call to list.files(..., full.names = TRUE)}
}
\value{
a big data.frame
}
\description{
reads a binary file in clocklab format
}
\examples{
f <- file.path(system.file(package = 'actogrammr'), 'testdata')
d <- read_clock_lab_file(file_name = list.files(path = f, full.names = TRUE)[1])
}
