% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clock_lab.R
\name{read_clock_lab_files}
\alias{read_clock_lab_files}
\title{read_clock_lab_files}
\usage{
read_clock_lab_files(file_names)
}
\arguments{
\item{file_names}{the names of the files to read.
Should be the result of a call to list.files(..., full.names = TRUE)}
}
\value{
a big data.frame
}
\description{
reads binary files in clocklab format
}
\examples{
f <- file.path(system.file(package = 'actogrammr'), 'testdata')
d <- read_clock_lab_files(file_names = list.files(path = f, full.names = TRUE))
}
