% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{bin_data}
\alias{bin_data}
\title{bin_data}
\usage{
bin_data(data, minutes_per_bin)
}
\arguments{
\item{data}{the activity data to bin}

\item{minutes_per_bin}{number of minutes per bin}
}
\value{
the data, after binning
}
