% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{bin_data}
\alias{bin_data}
\title{bin_data}
\usage{
bin_data(data, minutes_per_bin)
}
\arguments{
\item{data}{the activity data to bin}

\item{minutes_per_bin}{number of minutes per bin}
}
\value{
the data, after binning
}
\description{
function to bin data time-wise
}
\examples{
f <- file.path(system.file(package = 'actogrammr'), 'testdata')
d <- read_clock_lab_files(file_names = list.files(path = f, full.names = TRUE))
b <- bin_data(data = d, minutes_per_bin = 6)
}
