% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_actogram.R
\name{plot_actogram}
\alias{plot_actogram}
\title{plot_actogram}
\usage{
plot_actogram(data, start_date = min(data$date), end_date = max(data$date))
}
\arguments{
\item{data}{the activity data to plot}

\item{start_date}{the start time}

\item{end_date}{the end time}
}
\value{
the plot
}
\description{
plots an actogram
}
\examples{
f <- file.path(system.file(package = 'actogrammr'), 'testdata')
d <- read_clock_lab_files(file_names = list.files(path = f, full.names = TRUE))
b <- bin_data(data = d, minutes_per_bin = 6)
\dontrun{
plot_actogram(data = b, start_date = '2010-01-01')
}
}
